[Refactor] Move Paper class to knowledge module for better code organization
