+++ autoscholar/knowledge/__init__.py
@@ -1,0 +1,3 @@
+from autoscholar.knowledge.paper import Paper
+
+__all__ = ["Paper"]
